wip: moved image upload logic into separate file
