+++ pkg/deej/serial.go
@@ -5,15 +5,10 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"os"
 	"regexp"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
-
-	"github.com/ncruces/zenity"
 
 	"github.com/jacobsa/go-serial/serial"
 	"go.uber.org/zap"
@@ -35,8 +30,9 @@ type SerialIO struct {
 	connOptions serial.OpenOptions
 	conn        io.ReadWriteCloser
 
-	messageQueue  chan []byte
-	currentUpload *ImageUploadState
+	messageQueue       chan []byte
+	currentUpload      *ImageUploadState
+	currentMultiUpload *MultiUploadState
 
 	lastKnownNumSliders        int
 	lastKnownNumSwitches       int
@@ -55,15 +51,9 @@ type SliderMoveEvent struct {
 	PercentValue float32
 }
 
-type ImageUploadState struct {
-	Lock           sync.Mutex
-	LoadedFile     []byte
-	TotalBytesSent int
-	Dialog         *zenity.ProgressDialog
-}
-
 var expectedControlInput = regexp.MustCompile(`^\d{1,4}(\|\d{1,4})* \d(\|\d)*\r\n$`)
 var transferInput = regexp.MustCompile(`^OK(?: (?:READY|DONE)? ?\d+| \d+)\r\n$`)
+var transferMultipleInput = regexp.MustCompile(`^OK\s+FRAMES?\s+(READY|DONE)?\s*\d+\r\n$`)
 var transferFail = regexp.MustCompile(`^FAIL`)
 
 const UploadBlock = 1024
@@ -272,6 +262,14 @@ func (sio *SerialIO) handleLine(logger *zap.SugaredLogger, line string) {
 		return
 	}
 
+	// For when uploading image frames
+	if transferMultipleInput.MatchString(line) {
+		err := sio.transferAnimated(logger, line)
+		if err != nil {
+			logger.Errorw("Error when transfering animated image!", "error", err)
+		}
+	}
+
 	// Handle I/O errors
 	if transferFail.MatchString(line) {
 		sio.handleTransferError(logger, line)
@@ -293,6 +291,7 @@ func (sio *SerialIO) handleLine(logger *zap.SugaredLogger, line string) {
 	sio.handleSliders(logger, splitLine[0])
 	sio.handleSwitches(logger, splitLine[1])
 }
+
 
 func (sio *SerialIO) handleSwitches(logger *zap.SugaredLogger, line string) {
 	// split on pipe (|), this gives a slice of numerical strings between "0" and "1023"
@@ -423,106 +422,3 @@ func (sio *SerialIO) writeLine(logger *zap.SugaredLogger) {
 	}()
 }
 
-func (sio *SerialIO) StartImageUpload(logger *zap.SugaredLogger, path string) error {
-	sio.currentUpload = &ImageUploadState{}
-
-	sio.currentUpload.Lock.Lock()
-	defer sio.currentUpload.Lock.Unlock()
-
-	logger.Debugw("Attempting to open image", "path", path)
-	file, err := os.Open(path)
-	if err != nil {
-		return err
-	}
-
-	logger.Debug("Attempting to read bytes from file")
-	dat, err := ioutil.ReadAll(file)
-	if err != nil {
-		return err
-	}
-
-	logger.Debugw("Creating progress dialog")
-	dlg, err := zenity.Progress(zenity.Title("Uploading Image"), zenity.NoCancel())
-	if err != nil {
-		return err
-	}
-
-	err = dlg.Text(fmt.Sprintf("0/%d", len(dat)))
-	if err != nil {
-		return err
-	}
-
-	sio.currentUpload.LoadedFile = dat
-	sio.currentUpload.TotalBytesSent = 0
-	sio.currentUpload.Dialog = &dlg
-
-	sio.messageQueue <- []byte(fmt.Sprintf("sendimg %d", len(dat)))
-
-	return nil
-}
-
-func (sio *SerialIO) transferBlock(logger *zap.SugaredLogger, line string) error {
-	cu := sio.currentUpload
-	if cu == nil {
-		return errors.New("current upload object is empty")
-	}
-
-	logger.Debugw("Parsing line ", "line", line)
-
-	cu.Lock.Lock()
-	defer cu.Lock.Unlock()
-
-	dlg := *cu.Dialog
-
-	if cu.TotalBytesSent >= len(cu.LoadedFile) {
-		err := dlg.Complete()
-
-		if err != nil {
-			logger.Warnw("Issues cleaning up progress dialog", "error", err)
-		}
-		logger.Info("Upload completed")
-
-		return nil
-	}
-
-	end := util.MinInt(cu.TotalBytesSent+UploadBlock, len(cu.LoadedFile))
-	sio.messageQueue <- cu.LoadedFile[cu.TotalBytesSent:end]
-	cu.TotalBytesSent = end
-
-	err := dlg.Value(int((float64(cu.TotalBytesSent) / float64(len(cu.LoadedFile))) * 100))
-	err = dlg.Text(fmt.Sprintf("Uploaded %d/%d bytes", cu.TotalBytesSent, len(cu.LoadedFile)))
-
-	if err != nil {
-		logger.Warnw("Issues updating progress dialog", "error", err)
-	}
-
-	logger.Debugw("Sent bytes to controller", "sent", cu.TotalBytesSent, "total", len(cu.LoadedFile))
-
-	return nil
-}
-
-func (sio *SerialIO) handleTransferError(logger *zap.SugaredLogger, line string) {
-	cus := sio.currentUpload
-
-	cleaned := strings.TrimSuffix(line, "\r\n")
-	split := strings.Split(cleaned, ",")
-
-	reason := "Unknown"
-	if len(split) > 1 {
-		reason = split[1]
-	} else {
-		logger.Warnw("Controller did not give a reason for failure", "line", split)
-	}
-
-	logger.Errorw("Failed to transfer image to microcontroller", "reason", reason)
-
-	cus.Lock.Lock()
-	defer cus.Lock.Unlock()
-
-	cus.LoadedFile = []byte{}
-	cus.TotalBytesSent = 0
-
-	dlg := *cus.Dialog
-	dlg.Text(fmt.Sprintf("Transfer FAILED! Reason: %s", reason))
-	dlg.Complete()
-}
